+++ EBU_packets/EBUanalogOut.c
@@ -19,6 +19,6 @@ float getAnalogOut(EBUanalogOut *packet, int channel){
 	
 	uint16_t intValue = packet->channel[channel];
 	
-	float value = ((intValue/65535) * 5);
+	float value = (((float)intValue/65535) * 5);
 	return value;
 }
